Removed Topic Map code
The topic map code wasn't working as expected. Calling dec.getModel().get('trust_level') returns the trust level value of the topic author twice rather than the users listed in the topic map. As a proof of concept, dec.attrs.username does return the users listed in the topic map, but trust_level isn't an attribute unfortunately. Also this may be pulling those values from the posts rather than the topic map. More investigation will be need as far as that is concerned. I'm hoping to figure out what it would take to get trust level added as an available attribute, but thankfully the topic map isn't as prominent as the rest of the areas of flair.